Enemies diying animation
Now entites have a boolean to_die, probably the best option to kill "dynamic" entities such as player or enemies (so they do animations and dont die instantly). As for now, to the buff manager only can enter these two types. Can't think of more that could fit in it.
Feel free to go back to "to_delete" if it causes any trouble